separate asset module
