+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "python.linting.pylintEnabled": false,
+    "rust.features": ["bedrock/VK_EXT_debug_report"]
+}
